requirements.txt
torch>=2.5.1
torchvision>=0.20.1
ultralytics>=8.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.23.0
supervision>=0.15.0
pycocotools>=2.0.7
scikit-image>=0.21.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0